Foreign.JavaScript.TH

Documentation

class Monad m => HasJSContext m where

Minimal complete definition

askJSContext

Associated Types

type JSContextPhantom m :: *

Methods

askJSContext :: m (JSContextSingleton (JSContextPhantom m))

type HasWebView = HasJSContext

newtype WithJSContextSingleton x m a

Constructors

WithJSContextSingleton

Fields

unWithJSContextSingleton :: ReaderT (JSContextSingleton x) m a

runWithJSContextSingleton :: WithJSContextSingleton x m a -> JSContextSingleton x -> m a

withJSContextSingleton :: MonadJSM m => (forall x. JSContextSingleton x -> m r) -> m r

withJSContextSingletonMono :: MonadJSM m => (JSContextSingleton () -> m r) -> m r

Warning: withJSContextSingletonMono does not provide the same guarantees that withJSContextSingleton does.

newtype JSContextSingleton x

A singleton type for a given JSContext; we use this to statically guarantee that different JSContexts don't get mixed up

Constructors

JSContextSingleton

Fields

unJSContextSingleton :: JSContextRef

newtype JSFFI

Constructors

JSFFI String

data JSFun x

Constructors

JSFun

Fields

unJSFun :: JSRef x

unJSFunction :: Function

class IsJSContext x

Associated Types

data JSRef x

class (Monad m, MonadJSM (JSX m), MonadFix (JSX m), MonadJS x (JSX m)) => HasJS x m | m -> x where

Minimal complete definition

liftJS

Associated Types

type JSX m :: * -> *

Methods

liftJS :: JSX m a -> m a

class Monad m => MonadJS x m | m -> x where

A Monad that is capable of executing JavaScript

Minimal complete definition

runJS, forkJS, mkJSUndefined, isJSNull, isJSUndefined, fromJSBool, fromJSString, fromJSArray, fromJSUint8Array, fromJSNumber, withJSBool, withJSString, withJSNumber, withJSArray, withJSUint8Array, mkJSFun, freeJSFun, setJSProp, getJSProp, withJSNode

Methods

runJS :: JSFFI -> [JSRef x] -> m (JSRef x)

forkJS :: m () -> m ThreadId

mkJSUndefined :: m (JSRef x)

isJSNull :: JSRef x -> m Bool

isJSUndefined :: JSRef x -> m Bool

fromJSBool :: JSRef x -> m Bool

fromJSString :: JSRef x -> m String

fromJSArray :: JSRef x -> m [JSRef x]

fromJSUint8Array :: JSRef x -> m ByteString

fromJSNumber :: JSRef x -> m Double

withJSBool :: Bool -> (JSRef x -> m r) -> m r

withJSString :: String -> (JSRef x -> m r) -> m r

withJSNumber :: Double -> (JSRef x -> m r) -> m r

withJSArray :: [JSRef x] -> (JSRef x -> m r) -> m r

withJSUint8Array :: ByteString -> (JSUint8Array x -> m r) -> m r

mkJSFun :: ([JSRef x] -> m (JSRef x)) -> m (JSFun x)

Create a JSFun with zero arguments; should be equilvant to `syncCallback AlwaysRetain True` in GHCJS

freeJSFun :: JSFun x -> m ()

setJSProp :: String -> JSRef x -> JSRef x -> m ()

getJSProp :: String -> JSRef x -> m (JSRef x)

withJSNode :: Node -> (JSRef x -> m r) -> m r

data JSCtx_JavaScriptCore x

type HasJS' = HasJS (JSCtx_JavaScriptCore ())

newtype WithJSContext x m a

Constructors

WithJSContext

Fields

unWithJSContext :: ReaderT JSContextRef m a

runWithJSContext :: WithJSContext x m a -> JSContextRef -> m a

lowerWithJSContext :: MonadJSM m => WithJSContext x IO a -> m a

liftWithJSContextSingletonThroughWithJSContext :: (HasJSContext m, MonadJSM m, MonadTrans t, Monad m1) => ((t1 -> t m1 a) -> WithJSContext x IO b) -> (t1 -> WithJSContextSingleton (JSContextPhantom m) m1 a) -> m b

class FromJS x a where

Minimal complete definition

fromJS

Methods

fromJS :: MonadJS x m => JSRef x -> m a

class ToJS x a where

Minimal complete definition

withJS

Methods

withJS :: MonadJS x m => a -> (JSRef x -> m r) -> m r

newtype JSArray a

Constructors

JSArray

Fields

unJSArray :: [a]

withJSArrayFromList :: (ToJS x a, MonadJS x m) => [a] -> (JSRef x -> m r) -> m r

newtype JSUint8Array x

Constructors

JSUint8Array

Fields

unJSUint8Array :: JSRef x

importJS :: Safety -> String -> String -> Q Type -> Q [Dec]

mkJSFFI :: Safety -> String -> Q ([Dec], Exp)

parseType :: Type -> ([Type], Type)

data Safety :: *

Constructors

Unsafe

Safe

Interruptible