Reflex.Dom.Builder.Class

Documentation

class Default (EventSpec d EventResult) => DomSpace d where

Minimal complete definition

addEventSpecFlags

Associated Types

type EventSpec d :: (EventTag -> *) -> *

type RawTextNode d :: *

type RawElement d :: *

type RawFile d :: *

type RawInputElement d :: *

type RawTextAreaElement d :: *

type RawSelectElement d :: *

Methods

addEventSpecFlags :: proxy d -> EventName en -> (Maybe (er en) -> EventFlags) -> EventSpec d er -> EventSpec d er

class (Monad m, Reflex t, DomSpace (DomBuilderSpace m), NotReady t m, Adjustable t m) => DomBuilder t m | m -> t where

DomBuilder t m indicates that m is a Monad capable of building dynamic DOM in the Reflex timeline t

Associated Types

type DomBuilderSpace m :: *

Methods

textNode :: TextNodeConfig t -> m (TextNode (DomBuilderSpace m) t)

textNode :: (MonadTrans f, m ~ f m', DomBuilderSpace m' ~ DomBuilderSpace m, DomBuilder t m') => TextNodeConfig t -> m (TextNode (DomBuilderSpace m) t)

element :: Text -> ElementConfig er t (DomBuilderSpace m) -> m a -> m (Element er (DomBuilderSpace m) t, a)

element :: (MonadTransControl f, StT f a ~ a, m ~ f m', DomBuilderSpace m' ~ DomBuilderSpace m, DomBuilder t m') => Text -> ElementConfig er t (DomBuilderSpace m) -> m a -> m (Element er (DomBuilderSpace m) t, a)

inputElement :: InputElementConfig er t (DomBuilderSpace m) -> m (InputElement er (DomBuilderSpace m) t)

inputElement :: (MonadTransControl f, m ~ f m', DomBuilderSpace m' ~ DomBuilderSpace m, DomBuilder t m') => InputElementConfig er t (DomBuilderSpace m) -> m (InputElement er (DomBuilderSpace m) t)

textAreaElement :: TextAreaElementConfig er t (DomBuilderSpace m) -> m (TextAreaElement er (DomBuilderSpace m) t)

textAreaElement :: (MonadTransControl f, m ~ f m', DomBuilderSpace m' ~ DomBuilderSpace m, DomBuilder t m') => TextAreaElementConfig er t (DomBuilderSpace m) -> m (TextAreaElement er (DomBuilderSpace m) t)

selectElement :: SelectElementConfig er t (DomBuilderSpace m) -> m a -> m (SelectElement er (DomBuilderSpace m) t, a)

selectElement :: (MonadTransControl f, StT f a ~ a, m ~ f m', DomBuilderSpace m' ~ DomBuilderSpace m, DomBuilder t m') => SelectElementConfig er t (DomBuilderSpace m) -> m a -> m (SelectElement er (DomBuilderSpace m) t, a)

placeRawElement :: RawElement (DomBuilderSpace m) -> m ()

placeRawElement :: (MonadTrans f, m ~ f m', DomBuilderSpace m' ~ DomBuilderSpace m, DomBuilder t m') => RawElement (DomBuilderSpace m) -> m ()

wrapRawElement :: RawElement (DomBuilderSpace m) -> RawElementConfig er t (DomBuilderSpace m) -> m (Element er (DomBuilderSpace m) t)

wrapRawElement :: (MonadTrans f, m ~ f m', DomBuilderSpace m' ~ DomBuilderSpace m, DomBuilder t m') => RawElement (DomBuilderSpace m) -> RawElementConfig er t (DomBuilderSpace m) -> m (Element er (DomBuilderSpace m) t)

class DomBuilder t m => MountableDomBuilder t m where

Minimal complete definition

buildDomFragment, mountDomFragment

Associated Types

type DomFragment m :: *

Methods

buildDomFragment :: m a -> m (DomFragment m, a)

mountDomFragment :: DomFragment m -> Event t (DomFragment m) -> m ()

type Namespace = Text

data TextNodeConfig t

Constructors

TextNodeConfig

Fields

_textNodeConfig_initialContents :: !Text

_textNodeConfig_setContents :: !(Maybe (Event t Text))

newtype TextNode d t

Constructors

TextNode

Fields

_textNode_raw :: RawTextNode d

data AttributeName

Constructors

AttributeName !(Maybe Namespace) !Text

mapKeysToAttributeName :: Map Text v -> Map AttributeName v

data Propagation

Constructors

Propagation_Continue

Propagation_Stop

Propagation_StopImmediate

data EventFlags

Constructors

EventFlags

Fields

_eventFlags_propagation :: Propagation

_eventFlags_preventDefault :: Bool

preventDefault :: EventFlags

stopPropagation :: EventFlags

data ElementConfig er t s

Constructors

ElementConfig

Fields

_elementConfig_namespace :: Maybe Namespace

_elementConfig_initialAttributes :: Map AttributeName Text

_elementConfig_modifyAttributes :: Maybe (Event t (Map AttributeName (Maybe Text)))

_elementConfig_eventSpec :: EventSpec s er

data Element er d t

Constructors

Element

Fields

_element_events :: EventSelector t (WrapArg er EventName)

_element_raw :: RawElement d

data InputElementConfig er t s

Constructors

InputElementConfig

Fields

_inputElementConfig_initialValue :: Text

_inputElementConfig_setValue :: Maybe (Event t Text)

_inputElementConfig_initialChecked :: Bool

_inputElementConfig_setChecked :: Maybe (Event t Bool)

_inputElementConfig_elementConfig :: ElementConfig er t s

data InputElement er d t

Constructors

InputElement

Fields

_inputElement_value :: Dynamic t Text

_inputElement_checked :: Dynamic t Bool

_inputElement_checkedChange :: Event t Bool

_inputElement_input :: Event t Text

_inputElement_hasFocus :: Dynamic t Bool

_inputElement_element :: Element er d t

_inputElement_raw :: RawInputElement d

_inputElement_files :: Dynamic t [RawFile d]

data TextAreaElementConfig er t m

Constructors

TextAreaElementConfig

Fields

_textAreaElementConfig_initialValue :: Text

_textAreaElementConfig_setValue :: Maybe (Event t Text)

_textAreaElementConfig_elementConfig :: ElementConfig er t m

data TextAreaElement er d t

Constructors

TextAreaElement

Fields

_textAreaElement_value :: Dynamic t Text

_textAreaElement_input :: Event t Text

_textAreaElement_hasFocus :: Dynamic t Bool

_textAreaElement_element :: Element er d t

_textAreaElement_raw :: RawTextAreaElement d

extractRawElementConfig :: ElementConfig er t m -> RawElementConfig er t m

data RawElementConfig er t s

Constructors

RawElementConfig

Fields

_rawElementConfig_modifyAttributes :: Maybe (Event t (Map AttributeName (Maybe Text)))

_rawElementConfig_eventSpec :: EventSpec s er

data SelectElementConfig er t m

Constructors

SelectElementConfig

Fields

_selectElementConfig_initialValue :: Text

_selectElementConfig_setValue :: Maybe (Event t Text)

_selectElementConfig_elementConfig :: ElementConfig er t m

data SelectElement er d t

Constructors

SelectElement

Fields

_selectElement_element :: Element er d t

_selectElement_value :: Dynamic t Text

_selectElement_change :: Event t Text

Fires when the value is changed by the user, but not when it is set by setValue

_selectElement_hasFocus :: Dynamic t Bool

_selectElement_raw :: RawSelectElement d

selectElementConfig_initialValue :: forall er t m. Lens' (SelectElementConfig er t m) Text

selectElementConfig_elementConfig :: forall er t m er m. Lens (SelectElementConfig er t m) (SelectElementConfig er t m) (ElementConfig er t m) (ElementConfig er t m)

textAreaElementConfig_initialValue :: forall er t m. Lens' (TextAreaElementConfig er t m) Text

textAreaElementConfig_elementConfig :: forall er t m er m. Lens (TextAreaElementConfig er t m) (TextAreaElementConfig er t m) (ElementConfig er t m) (ElementConfig er t m)

elementConfig_namespace :: forall er t s. Lens' (ElementConfig er t s) (Maybe Namespace)

elementConfig_initialAttributes :: forall er t s. Lens' (ElementConfig er t s) (Map AttributeName Text)

elementConfig_eventSpec :: forall er t s er s. Lens (ElementConfig er t s) (ElementConfig er t s) (EventSpec s er) (EventSpec s er)

rawElementConfig_eventSpec :: forall er t s er s. Lens (RawElementConfig er t s) (RawElementConfig er t s) (EventSpec s er) (EventSpec s er)

inputElementConfig_initialValue :: forall er t s. Lens' (InputElementConfig er t s) Text

inputElementConfig_initialChecked :: forall er t s. Lens' (InputElementConfig er t s) Bool

inputElementConfig_elementConfig :: forall er t s er s. Lens (InputElementConfig er t s) (InputElementConfig er t s) (ElementConfig er t s) (ElementConfig er t s)

textNodeConfig_initialContents :: forall t. Lens' (TextNodeConfig t) Text

textNodeConfig_setContents :: Reflex t => Lens (TextNodeConfig t) (TextNodeConfig t) (Event t Text) (Event t Text)

This lens is technically illegal. The implementation of TextNodeConfig uses a Maybe under the hood for efficiency reasons. However, always interacting with TextNodeConfig via lenses will always behave correctly, and if you pattern match on it, you should always treat Nothing as never.

inputElementConfig_setValue :: Reflex t => Lens (InputElementConfig er t m) (InputElementConfig er t m) (Event t Text) (Event t Text)

This lens is technically illegal. The implementation of InputElementConfig uses a Maybe under the hood for efficiency reasons. However, always interacting with InputElementConfig via lenses will always behave correctly, and if you pattern match on it, you should always treat Nothing as never.

inputElementConfig_setChecked :: Reflex t => Lens (InputElementConfig er t m) (InputElementConfig er t m) (Event t Bool) (Event t Bool)

This lens is technically illegal. The implementation of InputElementConfig uses a Maybe under the hood for efficiency reasons. However, always interacting with InputElementConfig via lenses will always behave correctly, and if you pattern match on it, you should always treat Nothing as never.

rawElementConfig_modifyAttributes :: Reflex t => Lens (RawElementConfig er t m) (RawElementConfig er t m) (Event t (Map AttributeName (Maybe Text))) (Event t (Map AttributeName (Maybe Text)))

This lens is technically illegal. The implementation of RawElementConfig uses a Maybe under the hood for efficiency reasons. However, always interacting with RawElementConfig via lenses will always behave correctly, and if you pattern match on it, you should always treat Nothing as never.

elementConfig_modifyAttributes :: Reflex t => Lens (ElementConfig er t m) (ElementConfig er t m) (Event t (Map AttributeName (Maybe Text))) (Event t (Map AttributeName (Maybe Text)))

This lens is technically illegal. The implementation of RawElementConfig uses a Maybe under the hood for efficiency reasons. However, always interacting with RawElementConfig via lenses will always behave correctly, and if you pattern match on it, you should always treat Nothing as never.

textAreaElementConfig_setValue :: Reflex t => Lens (TextAreaElementConfig er t m) (TextAreaElementConfig er t m) (Event t Text) (Event t Text)

This lens is technically illegal. The implementation of TextAreaElementConfig uses a Maybe under the hood for efficiency reasons. However, always interacting with TextAreaElementConfig via lenses will always behave correctly, and if you pattern match on it, you should always treat Nothing as never.

selectElementConfig_setValue :: Reflex t => Lens (SelectElementConfig er t m) (SelectElementConfig er t m) (Event t Text) (Event t Text)

This lens is technically illegal. The implementation of SelectElementConfig uses a Maybe under the hood for efficiency reasons. However, always interacting with SelectElementConfig via lenses will always behave correctly, and if you pattern match on it, you should always treat Nothing as never.

class InitialAttributes a where

Minimal complete definition

initialAttributes

Methods

initialAttributes :: Lens' a (Map AttributeName Text)

class ModifyAttributes t a | a -> t where

Minimal complete definition

modifyAttributes

Methods

modifyAttributes :: Reflex t => Lens' a (Event t (Map AttributeName (Maybe Text)))

class HasNamespace a where

Minimal complete definition

namespace

Methods

namespace :: Lens' a (Maybe Namespace)

Convenience functions

class HasDomEvent t target eventName where

Minimal complete definition

domEvent

Associated Types

type DomEventType target eventName :: *

Methods

domEvent :: EventName eventName -> target -> Event t (DomEventType target eventName)

type LiftDomBuilder t f m = (Reflex t, MonadTransControlStateless f, Monad m, DomBuilder t m, DomBuilderSpace (f m) ~ DomBuilderSpace m)

class MonadTransControl t => MonadTransControlStateless t where

Methods

stTCoercion :: proxy t -> Coercion (StT t a) a

stTCoercion :: a ~ StT t a => proxy t -> Coercion (StT t a) a

toStT :: MonadTransControlStateless t => proxy t -> a -> StT t a

fromStT :: MonadTransControlStateless t => proxy t -> StT t a -> a

type RunStateless t = forall n b. Monad n => t n b -> n b

liftWithStateless :: forall m t a. (Monad m, MonadTransControlStateless t) => (RunStateless t -> m a) -> t m a

liftTextNode :: (MonadTrans f, DomBuilder t m) => TextNodeConfig t -> f m (TextNode (DomBuilderSpace m) t)

liftElement :: LiftDomBuilder t f m => Text -> ElementConfig er t (DomBuilderSpace m) -> f m a -> f m (Element er (DomBuilderSpace m) t, a)

class (Reflex t, Monad m) => DomRenderHook t m | m -> t where

Minimal complete definition

withRenderHook, requestDomAction, requestDomAction_

Methods

withRenderHook :: (forall x. JSM x -> JSM x) -> m a -> m a

requestDomAction :: Event t (JSM a) -> m (Event t a)

requestDomAction_ :: Event t (JSM a) -> m ()

liftElementConfig :: ElementConfig er t s -> ElementConfig er t s

Deprecated: Use id instead; this function is no longer necessary

module Reflex.Dom.Builder.Class.Events

module Reflex.NotReady.Class