Reflex.Dom.Builder.Immediate

Documentation

newtype EventTriggerRef t a :: * -> * -> *

A reference to an EventTrigger suitable for firing with TriggerEventT.

Constructors

EventTriggerRef

Fields

unEventTriggerRef :: IORef (Maybe (EventTrigger t a))

data ImmediateDomBuilderEnv t

Constructors

ImmediateDomBuilderEnv

Fields

_immediateDomBuilderEnv_document :: !Document

_immediateDomBuilderEnv_parent :: !Node

_immediateDomBuilderEnv_unreadyChildren :: !(IORef Word)

_immediateDomBuilderEnv_commitAction :: !(JSM ())

newtype ImmediateDomBuilderT t m a

Constructors

ImmediateDomBuilderT

Fields

unImmediateDomBuilderT :: ReaderT (ImmediateDomBuilderEnv t) (RequesterT t JSM Identity (TriggerEventT t m)) a

runImmediateDomBuilderT :: (MonadFix m, PerformEvent t m, MonadReflexCreateTrigger t m, MonadJSM m, MonadJSM (Performable m), MonadRef m, Ref m ~ IORef) => ImmediateDomBuilderT t m a -> ImmediateDomBuilderEnv t -> Chan [DSum (EventTriggerRef t) TriggerInvocation] -> m a

askParent :: Monad m => ImmediateDomBuilderT t m Node

askEvents :: Monad m => ImmediateDomBuilderT t m (Chan [DSum (EventTriggerRef t) TriggerInvocation])

append :: MonadJSM m => Node -> ImmediateDomBuilderT t m ()

textNodeInternal :: (MonadJSM m, ToDOMString contents) => contents -> ImmediateDomBuilderT t m Text

deleteBetweenExclusive :: (MonadJSM m, IsNode start, IsNode end) => start -> end -> m ()

s and e must both be children of the same node and s must precede e; all nodes between s and e will be removed, but s and e will not be removed

extractBetweenExclusive :: (MonadJSM m, IsNode start, IsNode end) => DocumentFragment -> start -> end -> m ()

s and e must both be children of the same node and s must precede e; all nodes between s and e will be moved into the given DocumentFragment, but s and e will not be moved

deleteUpTo :: (MonadJSM m, IsNode start, IsNode end) => start -> end -> m ()

s and e must both be children of the same node and s must precede e; s and all nodes between s and e will be removed, but e will not be removed

extractUpTo :: (MonadJSM m, IsNode start, IsNode end) => DocumentFragment -> start -> end -> m ()

type SupportsImmediateDomBuilder t m = (Reflex t, MonadJSM m, MonadHold t m, MonadFix m, MonadReflexCreateTrigger t m, MonadRef m, Ref m ~ Ref JSM, Adjustable t m, PrimMonad m)

collectUpTo :: (MonadJSM m, IsNode start, IsNode end) => start -> end -> m DocumentFragment

collectUpToGivenParent :: (MonadJSM m, IsNode parent, IsNode start, IsNode end) => parent -> start -> end -> m DocumentFragment

newtype EventFilterTriggerRef t er en

Constructors

EventFilterTriggerRef (IORef (Maybe (EventTrigger t (er en))))

wrap :: forall m er t. (Reflex t, MonadFix m, MonadJSM m, MonadReflexCreateTrigger t m) => RawElement GhcjsDomSpace -> RawElementConfig er t GhcjsDomSpace -> ImmediateDomBuilderT t m (Element er GhcjsDomSpace t)

makeElement :: forall er t m a. (MonadJSM m, MonadFix m, MonadReflexCreateTrigger t m, Adjustable t m) => Text -> ElementConfig er t GhcjsDomSpace -> ImmediateDomBuilderT t m a -> ImmediateDomBuilderT t m ((Element er GhcjsDomSpace t, a), Element)

newtype GhcjsDomHandler a b

Constructors

GhcjsDomHandler

Fields

unGhcjsDomHandler :: a -> JSM b

newtype GhcjsDomHandler1 a b

Constructors

GhcjsDomHandler1

Fields

unGhcjsDomHandler1 :: forall x. a x -> JSM (b x)

newtype GhcjsDomEvent en

Constructors

GhcjsDomEvent

Fields

unGhcjsDomEvent :: EventType en

data GhcjsDomSpace

newtype GhcjsEventFilter er en

Constructors

GhcjsEventFilter (GhcjsDomEvent en -> JSM (EventFlags, JSM (Maybe (er en))))

data Pair1 f g a

Constructors

Pair1 (f a) (g a)

data Maybe1 f a

Constructors

Nothing1

Just1 (f a)

data GhcjsEventSpec er

Constructors

GhcjsEventSpec

Fields

_ghcjsEventSpec_filters :: DMap EventName (GhcjsEventFilter er)

_ghcjsEventSpec_handler :: GhcjsEventHandler er

class Monad m => HasDocument m where

Methods

askDocument :: m Document

askDocument :: (m ~ f m', MonadTrans f, Monad m', HasDocument m') => m Document

ghcjsEventSpec_filters :: forall er. Lens' (GhcjsEventSpec er) (DMap EventName (GhcjsEventFilter er))

ghcjsEventSpec_handler :: forall er. Lens' (GhcjsEventSpec er) (GhcjsEventHandler er)

newtype GhcjsEventHandler er

Constructors

GhcjsEventHandler

Fields

unGhcjsEventHandler :: forall en. (EventName en, GhcjsDomEvent en) -> JSM (Maybe (er en))

drawChildUpdate :: (MonadIO m, MonadJSM m) => ImmediateDomBuilderEnv t -> (IORef (ChildReadyState k) -> JSM ()) -> ImmediateDomBuilderT t m (v' a) -> RequesterT t JSM Identity (TriggerEventT t m) (Compose ((,,,) DocumentFragment Text (IORef (ChildReadyState k))) v' a)

data ChildReadyState k

Constructors

ChildReadyState_Ready

ChildReadyState_Unready !(Maybe (Some k))

data ChildReadyStateInt

Constructors

ChildReadyStateInt_Ready

ChildReadyStateInt_Unready !(Maybe Int)

mkHasFocus :: (MonadHold t m, Reflex t) => Element er d t -> m (Dynamic t Bool)

insertBefore :: (MonadJSM m, IsNode new, IsNode existing) => new -> existing -> m ()

type family EventType en where ...

Equations

EventType AbortTag = UIEvent

EventType BlurTag = FocusEvent

EventType ChangeTag = Event

EventType ClickTag = MouseEvent

EventType ContextmenuTag = MouseEvent

EventType DblclickTag = MouseEvent

EventType DragTag = MouseEvent

EventType DragendTag = MouseEvent

EventType DragenterTag = MouseEvent

EventType DragleaveTag = MouseEvent

EventType DragoverTag = MouseEvent

EventType DragstartTag = MouseEvent

EventType DropTag = MouseEvent

EventType ErrorTag = UIEvent

EventType FocusTag = FocusEvent

EventType InputTag = Event

EventType InvalidTag = Event

EventType KeydownTag = KeyboardEvent

EventType KeypressTag = KeyboardEvent

EventType KeyupTag = KeyboardEvent

EventType LoadTag = UIEvent

EventType MousedownTag = MouseEvent

EventType MouseenterTag = MouseEvent

EventType MouseleaveTag = MouseEvent

EventType MousemoveTag = MouseEvent

EventType MouseoutTag = MouseEvent

EventType MouseoverTag = MouseEvent

EventType MouseupTag = MouseEvent

EventType MousewheelTag = MouseEvent

EventType ScrollTag = UIEvent

EventType SelectTag = UIEvent

EventType SubmitTag = Event

EventType WheelTag = WheelEvent

EventType BeforecutTag = ClipboardEvent

EventType CutTag = ClipboardEvent

EventType BeforecopyTag = ClipboardEvent

EventType CopyTag = ClipboardEvent

EventType BeforepasteTag = ClipboardEvent

EventType PasteTag = ClipboardEvent

EventType ResetTag = Event

EventType SearchTag = Event

EventType SelectstartTag = Event

EventType TouchstartTag = TouchEvent

EventType TouchmoveTag = TouchEvent

EventType TouchendTag = TouchEvent

EventType TouchcancelTag = TouchEvent

defaultDomEventHandler :: IsElement e => e -> EventName en -> EventM e (EventType en) (Maybe (EventResult en))

defaultDomWindowEventHandler :: Window -> EventName en -> EventM Window (EventType en) (Maybe (EventResult en))

withIsEvent :: EventName en -> (IsEvent (EventType en) => r) -> r

showEventName :: EventName en -> String

elementOnEventName :: IsElement e => EventName en -> e -> EventM e (EventType en) () -> JSM (JSM ())

windowOnEventName :: EventName en -> Window -> EventM Window (EventType en) () -> JSM (JSM ())

wrapDomEvent :: (TriggerEvent t m, MonadJSM m) => e -> (e -> EventM e event () -> JSM (JSM ())) -> EventM e event a -> m (Event t a)

subscribeDomEvent :: (EventM e event () -> JSM (JSM ())) -> EventM e event (Maybe a) -> Chan [DSum (EventTriggerRef t) TriggerInvocation] -> EventTrigger t a -> JSM (JSM ())

wrapDomEventMaybe :: (TriggerEvent t m, MonadJSM m) => e -> (e -> EventM e event () -> JSM (JSM ())) -> EventM e event (Maybe a) -> m (Event t a)

wrapDomEventsMaybe :: (MonadJSM m, MonadReflexCreateTrigger t m) => e -> (forall en. IsEvent (EventType en) => EventName en -> EventM e (EventType en) (Maybe (f en))) -> (forall en. EventName en -> e -> EventM e (EventType en) () -> JSM (JSM ())) -> ImmediateDomBuilderT t m (EventSelector t (WrapArg f EventName))

getKeyEvent :: EventM e KeyboardEvent Word

getMouseEventCoords :: EventM e MouseEvent (Int, Int)

getTouchEvent :: EventM e TouchEvent TouchEventResult

data WindowConfig t

Constructors

WindowConfig

data Window t

Constructors

Window

Fields

_window_events :: EventSelector t (WrapArg EventResult EventName)

_window_raw :: Window

wrapWindow :: (MonadJSM m, MonadReflexCreateTrigger t m) => Window -> WindowConfig t -> ImmediateDomBuilderT t m (Window t)

Internal

traverseDMapWithKeyWithAdjust' :: forall t m k v v'. (Adjustable t m, MonadHold t m, MonadFix m, MonadIO m, MonadJSM m, PrimMonad m, GCompare k) => (forall a. k a -> v a -> ImmediateDomBuilderT t m (v' a)) -> DMap k v -> Event t (PatchDMap k v) -> ImmediateDomBuilderT t m (DMap k v', Event t (PatchDMap k v'))

hoistTraverseWithKeyWithAdjust

Arguments

:: forall (k :: * -> *). (Adjustable t m, MonadHold t m, GCompare k, MonadIO m, MonadJSM m, PrimMonad m, MonadFix m, Patch (p k v), PatchTarget (p k (Constant Int)) ~ DMap k (Constant Int), Monoid (p k (Compose ((,,,) DocumentFragment Text (IORef (ChildReadyState k))) v')), Patch (p k (Constant Int)))

=> (forall vv vv'. (forall a. k a -> vv a -> RequesterT t JSM Identity (TriggerEventT t m) (vv' a)) -> DMap k vv -> Event t (p k vv) -> RequesterT t JSM Identity (TriggerEventT t m) (DMap k vv', Event t (p k vv')))
The base monad's traversal

-> (forall vv vv'. (forall a. vv a -> vv' a) -> p k vv -> p k vv')
A way of mapping over the patch type

-> (p k (Compose ((,,,) DocumentFragment Text (IORef (ChildReadyState k))) v') -> DMap k (Constant (IORef (ChildReadyState k))) -> IO (DMap k (Constant (IORef (ChildReadyState k)))))
Given a patch for the children DOM elements, produce a patch for the childrens' unreadiness state

-> (IORef (Map (Some k) Text) -> IORef Text -> p k (Compose ((,,,) DocumentFragment Text (IORef (ChildReadyState k))) v') -> JSM ())
Apply a patch to the DOM

-> (forall a. k a -> v a -> ImmediateDomBuilderT t m (v' a))

-> DMap k v

-> Event t (p k v)

-> ImmediateDomBuilderT t m (DMap k v', Event t (p k v'))

traverseIntMapWithKeyWithAdjust' :: forall t m v v'. (Adjustable t m, MonadHold t m, MonadFix m, MonadIO m, MonadJSM m, PrimMonad m) => (Key -> v -> ImmediateDomBuilderT t m v') -> IntMap v -> Event t (PatchIntMap v) -> ImmediateDomBuilderT t m (IntMap v', Event t (PatchIntMap v'))

hoistTraverseIntMapWithKeyWithAdjust

Arguments

:: (Adjustable t m, MonadIO m, MonadJSM m, MonadFix m, PrimMonad m, Monoid (p (DocumentFragment, Text, IORef ChildReadyStateInt, v')), Functor p)

=> ((Key -> v -> RequesterT t JSM Identity (TriggerEventT t m) (DocumentFragment, Text, IORef ChildReadyStateInt, v')) -> IntMap v -> Event t (p v) -> RequesterT t JSM Identity (TriggerEventT t m) (IntMap (DocumentFragment, Text, IORef ChildReadyStateInt, v'), Event t (p (DocumentFragment, Text, IORef ChildReadyStateInt, v'))))
The base monad's traversal

-> (p (DocumentFragment, Text, IORef ChildReadyStateInt, v') -> IntMap (IORef ChildReadyStateInt) -> IO (IntMap (IORef ChildReadyStateInt)))
Given a patch for the children DOM elements, produce a patch for the childrens' unreadiness state

-> (IORef (IntMap Text) -> IORef Text -> p (DocumentFragment, Text, IORef ChildReadyStateInt, v') -> JSM ())
Apply a patch to the DOM

-> (Key -> v -> ImmediateDomBuilderT t m v')

-> IntMap v

-> Event t (p v)

-> ImmediateDomBuilderT t m (IntMap v', Event t (p v'))