Reflex.Dom.Builder.Static

Documentation

data StaticDomBuilderEnv t

Constructors

StaticDomBuilderEnv

Fields

_staticDomBuilderEnv_shouldEscape :: Bool

_staticDomBuilderEnv_selectValue :: Maybe (Behavior t Text)

When the parent element is a "select" whose value has been set, this value tells us the current value. We use this to add a "selected" attribute to the appropriate "option" child element. This is not yet a perfect simulation of what the browser does, but it is much closer than doing nothing. TODO: Handle edge cases, e.g. setting to a value for which there is no option, then adding that option dynamically afterwards.

newtype StaticDomBuilderT t m a

Constructors

StaticDomBuilderT

Fields

unStaticDomBuilderT :: ReaderT (StaticDomBuilderEnv t) (StateT [Behavior t Builder] m) a

runStaticDomBuilderT :: (Monad m, Reflex t) => StaticDomBuilderT t m a -> StaticDomBuilderEnv t -> m (a, Behavior t Builder)

type SupportsStaticDomBuilder t m = (Reflex t, MonadIO m, MonadHold t m, MonadFix m, PerformEvent t m, MonadReflexCreateTrigger t m, MonadRef m, Ref m ~ Ref IO, Adjustable t m)

data StaticDomSpace

data StaticDomEvent a

Static documents never produce any events, so this type has no inhabitants

data StaticDomHandler a b

Static documents don't process events, so all handlers are equivalent

Constructors

StaticDomHandler

data StaticEventSpec er

Constructors

StaticEventSpec

hoistDMapWithKeyWithAdjust

Arguments

:: forall (k :: * -> *). (Adjustable t m, MonadHold t m, PatchTarget (p k (Constant (Behavior t Builder))) ~ DMap k (Constant (Behavior t Builder)), Patch (p k (Constant (Behavior t Builder))))

=> (forall vv vv'. (forall a. k a -> vv a -> m (vv' a)) -> DMap k vv -> Event t (p k vv) -> m (DMap k vv', Event t (p k vv')))
The base monad's traversal

-> (forall vv vv'. (forall a. vv a -> vv' a) -> p k vv -> p k vv')
A way of mapping over the patch type

-> (forall a. k a -> v a -> StaticDomBuilderT t m (v' a))

-> DMap k v

-> Event t (p k v)

-> StaticDomBuilderT t m (DMap k v', Event t (p k v'))

type StaticWidget x = PostBuildT Spider (StaticDomBuilderT Spider (PerformEventT Spider (SpiderHost Global)))

renderStatic :: StaticWidget x a -> IO (a, ByteString)