Reflex.Dom.Location

Documentation

browserHistoryWith

Arguments

:: (MonadJSM m, TriggerEvent t m, MonadHold t m)

=> (forall jsm. MonadJSM jsm => Location -> jsm a)
A function to encode the window location in a more useful form (e.g. getLocationAfterHost).

-> m (Dynamic t a)

Builds a Dynamic carrying the current window location.

getLocationAfterHost :: MonadJSM m => m Text

Returns the URI-decoded location after the host and port; i.e. returns the path, query, and fragment of the location.

getLocationFragment :: MonadJSM m => m Text

Returns the URI-decoded fragment/hash of the current window location.

getLocationHost :: MonadJSM m => m Text

Returns the host of the current window location

getLocationPath :: MonadJSM m => m Text

Returns the URI-decoded path of the current window location.

getLocationProtocol :: MonadJSM m => m Text

Returns the protocol/scheme (e.g. http: or https:) of the current window location

getLocationUrl :: MonadJSM m => m Text

Returns the full URI-decoded URL of the current window location.

manageHistory :: (MonadJSM m, TriggerEvent t m, MonadHold t m, PerformEvent t m, MonadJSM (Performable m)) => Event t HistoryCommand -> m (Dynamic t HistoryItem)

data HistoryCommand

Constructors

HistoryCommand_PushState HistoryStateUpdate

HistoryCommand_ReplaceState HistoryStateUpdate

data HistoryStateUpdate

Constructors

HistoryStateUpdate

Fields

_historyStateUpdate_state :: SerializedScriptValue

_historyStateUpdate_title :: Text

_historyStateUpdate_uri :: Maybe URI

If Just, update the URI; otherwise leave it unchanged NOTE: All URIs in this module are assumed to be already percent-escaped

data HistoryItem

Constructors

HistoryItem

Fields

_historyItem_state :: SerializedScriptValue

_historyItem_uri :: URI

NOTE: All URIs in this module are assumed to be already percent-escaped

getLocationUri :: MonadJSM m => Location -> m URI