Reflex.Dom.Widget.Input

Documentation

data TextInput t

Constructors

TextInput

Fields

_textInput_value :: Dynamic t Text

_textInput_input :: Event t Text

_textInput_keypress :: Event t Word

_textInput_keydown :: Event t Word

_textInput_keyup :: Event t Word

_textInput_hasFocus :: Dynamic t Bool

_textInput_builderElement :: InputElement EventResult GhcjsDomSpace t

_textInput_element :: TextInput t -> HTMLInputElement

data TextInputConfig t

Constructors

TextInputConfig

Fields

_textInputConfig_inputType :: Text

_textInputConfig_initialValue :: Text

_textInputConfig_setValue :: Event t Text

_textInputConfig_attributes :: Dynamic t (Map Text Text)

textInput :: (DomBuilder t m, PostBuild t m, DomBuilderSpace m ~ GhcjsDomSpace) => TextInputConfig t -> m (TextInput t)

Create an input whose value is a string. By default, the "type" attribute is set to "text", but it can be changed using the _textInputConfig_inputType field. Note that only types for which the value is always a string will work - types whose value may be null will not work properly with this widget.

textInputGetEnter :: Reflex t => TextInput t -> Event t ()

Deprecated: Use 'keypress Enter' instead

keypress :: (Reflex t, HasDomEvent t e KeypressTag, DomEventType e KeypressTag ~ Word) => Key -> e -> Event t ()

keydown :: (Reflex t, HasDomEvent t e KeydownTag, DomEventType e KeydownTag ~ Word) => Key -> e -> Event t ()

keyup :: (Reflex t, HasDomEvent t e KeyupTag, DomEventType e KeyupTag ~ Word) => Key -> e -> Event t ()

data RangeInputConfig t

Constructors

RangeInputConfig

Fields

_rangeInputConfig_initialValue :: Float

_rangeInputConfig_setValue :: Event t Float

_rangeInputConfig_attributes :: Dynamic t (Map Text Text)

data RangeInput t

Constructors

RangeInput

Fields

_rangeInput_value :: Dynamic t Float

_rangeInput_input :: Event t Float

_rangeInput_mouseup :: Event t (Int, Int)

_rangeInput_hasFocus :: Dynamic t Bool

_rangeInput_element :: HTMLInputElement

rangeInput :: (DomBuilder t m, PostBuild t m, DomBuilderSpace m ~ GhcjsDomSpace) => RangeInputConfig t -> m (RangeInput t)

Create an input whose value is a float. https://www.w3.org/wiki/HTML/Elements/input/range

data TextAreaConfig t

Constructors

TextAreaConfig

Fields

_textAreaConfig_initialValue :: Text

_textAreaConfig_setValue :: Event t Text

_textAreaConfig_attributes :: Dynamic t (Map Text Text)

data TextArea t

Constructors

TextArea

Fields

_textArea_value :: Dynamic t Text

_textArea_input :: Event t Text

_textArea_hasFocus :: Dynamic t Bool

_textArea_keypress :: Event t Word

_textArea_element :: HTMLTextAreaElement

textArea :: (DomBuilder t m, PostBuild t m, DomBuilderSpace m ~ GhcjsDomSpace) => TextAreaConfig t -> m (TextArea t)

data CheckboxConfig t

Constructors

CheckboxConfig

Fields

_checkboxConfig_setValue :: Event t Bool

_checkboxConfig_attributes :: Dynamic t (Map Text Text)

data Checkbox t

Constructors

Checkbox

Fields

_checkbox_value :: Dynamic t Bool

_checkbox_change :: Event t Bool

checkbox :: (DomBuilder t m, PostBuild t m) => Bool -> CheckboxConfig t -> m (Checkbox t)

Create an editable checkbox Note: if the "type" or "checked" attributes are provided as attributes, they will be ignored

type family CheckboxViewEventResultType (en :: EventTag) :: * where ...

Equations

CheckboxViewEventResultType ClickTag = Bool

CheckboxViewEventResultType t = EventResultType t

regularToCheckboxViewEventType :: EventName t -> EventResultType t -> CheckboxViewEventResultType t

newtype CheckboxViewEventResult en

Constructors

CheckboxViewEventResult

Fields

unCheckboxViewEventResult :: CheckboxViewEventResultType en

checkboxView :: forall t m. (DomBuilder t m, DomBuilderSpace m ~ GhcjsDomSpace, PostBuild t m, MonadHold t m) => Dynamic t (Map Text Text) -> Dynamic t Bool -> m (Event t Bool)

data FileInput d t

Constructors

FileInput

Fields

_fileInput_value :: Dynamic t [File]

_fileInput_element :: RawInputElement d

newtype FileInputConfig t

Constructors

FileInputConfig

Fields

_fileInputConfig_attributes :: Dynamic t (Map Text Text)

fileInput :: forall t m. (MonadIO m, MonadJSM m, MonadFix m, MonadHold t m, TriggerEvent t m, DomBuilder t m, PostBuild t m, DomBuilderSpace m ~ GhcjsDomSpace) => FileInputConfig t -> m (FileInput (DomBuilderSpace m) t)

data Dropdown t k

Constructors

Dropdown

Fields

_dropdown_value :: Dynamic t k

_dropdown_change :: Event t k

data DropdownConfig t k

Constructors

DropdownConfig

Fields

_dropdownConfig_setValue :: Event t k

_dropdownConfig_attributes :: Dynamic t (Map Text Text)

type family DropdownViewEventResultType (en :: EventTag) :: * where ...

Equations

DropdownViewEventResultType ChangeTag = Text

DropdownViewEventResultType t = EventResultType t

newtype DropdownViewEventResult en

Constructors

DropdownViewEventResult

Fields

unDropdownViewEventResult :: DropdownViewEventResultType en

regularToDropdownViewEventType :: EventName t -> EventResultType t -> DropdownViewEventResultType t

dropdown :: forall k t m. (DomBuilder t m, MonadFix m, MonadHold t m, PostBuild t m, Ord k) => k -> Dynamic t (Map k Text) -> DropdownConfig t k -> m (Dropdown t k)

Create a dropdown box The first argument gives the initial value of the dropdown; if it is not present in the map of options provided, it will be added with an empty string as its text

checkbox_value :: forall t. Lens' (Checkbox t) (Dynamic t Bool)

checkbox_change :: forall t. Lens' (Checkbox t) (Event t Bool)

checkboxConfig_setValue :: forall t. Lens' (CheckboxConfig t) (Event t Bool)

checkboxConfig_attributes :: forall t. Lens' (CheckboxConfig t) (Dynamic t (Map Text Text))

dropdown_value :: forall t k. Lens' (Dropdown t k) (Dynamic t k)

dropdown_change :: forall t k. Lens' (Dropdown t k) (Event t k)

dropdownConfig_setValue :: forall t k k. Lens (DropdownConfig t k) (DropdownConfig t k) (Event t k) (Event t k)

dropdownConfig_attributes :: forall t k. Lens' (DropdownConfig t k) (Dynamic t (Map Text Text))

fileInput_value :: forall d t t. Lens (FileInput d t) (FileInput d t) (Dynamic t [File]) (Dynamic t [File])

fileInput_element :: forall d t d. Lens (FileInput d t) (FileInput d t) (RawInputElement d) (RawInputElement d)

fileInputConfig_attributes :: forall t t. Iso (FileInputConfig t) (FileInputConfig t) (Dynamic t (Map Text Text)) (Dynamic t (Map Text Text))

rangeInput_value :: forall t. Lens' (RangeInput t) (Dynamic t Float)

rangeInput_mouseup :: forall t. Lens' (RangeInput t) (Event t (Int, Int))

rangeInput_input :: forall t. Lens' (RangeInput t) (Event t Float)

rangeInput_hasFocus :: forall t. Lens' (RangeInput t) (Dynamic t Bool)

rangeInput_element :: forall t. Lens' (RangeInput t) HTMLInputElement

rangeInputConfig_setValue :: forall t. Lens' (RangeInputConfig t) (Event t Float)

rangeInputConfig_initialValue :: forall t. Lens' (RangeInputConfig t) Float

rangeInputConfig_attributes :: forall t. Lens' (RangeInputConfig t) (Dynamic t (Map Text Text))

textInput_value :: forall t. Lens' (TextInput t) (Dynamic t Text)

textInput_keyup :: forall t. Lens' (TextInput t) (Event t Word)

textInput_keypress :: forall t. Lens' (TextInput t) (Event t Word)

textInput_keydown :: forall t. Lens' (TextInput t) (Event t Word)

textInput_input :: forall t. Lens' (TextInput t) (Event t Text)

textInput_hasFocus :: forall t. Lens' (TextInput t) (Dynamic t Bool)

textInput_builderElement :: forall t. Lens' (TextInput t) (InputElement EventResult GhcjsDomSpace t)

textInputConfig_setValue :: forall t. Lens' (TextInputConfig t) (Event t Text)

textInputConfig_inputType :: forall t. Lens' (TextInputConfig t) Text

textInputConfig_initialValue :: forall t. Lens' (TextInputConfig t) Text

textInputConfig_attributes :: forall t. Lens' (TextInputConfig t) (Dynamic t (Map Text Text))

textArea_value :: forall t. Lens' (TextArea t) (Dynamic t Text)

textArea_keypress :: forall t. Lens' (TextArea t) (Event t Word)

textArea_input :: forall t. Lens' (TextArea t) (Event t Text)

textArea_hasFocus :: forall t. Lens' (TextArea t) (Dynamic t Bool)

textArea_element :: forall t. Lens' (TextArea t) HTMLTextAreaElement

textAreaConfig_setValue :: forall t. Lens' (TextAreaConfig t) (Event t Text)

textAreaConfig_initialValue :: forall t. Lens' (TextAreaConfig t) Text

textAreaConfig_attributes :: forall t. Lens' (TextAreaConfig t) (Dynamic t (Map Text Text))

class HasSetValue a where

Minimal complete definition

setValue

Associated Types

type SetValue a :: *

Methods

setValue :: Lens' a (SetValue a)

class HasValue a where

Minimal complete definition

value

Associated Types

type Value a :: *

Methods

value :: a -> Value a

def :: Default a => a

The default value for this type.

(&) :: a -> (a -> b) -> b infixl 1

& is a reverse application operator. This provides notational convenience. Its precedence is one higher than that of the forward application operator $, which allows & to be nested in $.

Since: 4.8.0.0

(.~) :: ASetter s t a b -> b -> s -> t infixr 4

Replace the target of a Lens or all of the targets of a Setter or Traversal with a constant value.

This is an infix version of set, provided for consistency with (.=).

f <$ a ≡ mapped .~ f $ a

>>> (a,b,c,d) & _4 .~ e
(a,b,c,e)

>>> (42,"world") & _1 .~ "hello"
("hello","world")

>>> (a,b) & both .~ c
(c,c)

(.~) :: Setter s t a b    -> b -> s -> t
(.~) :: Iso s t a b       -> b -> s -> t
(.~) :: Lens s t a b      -> b -> s -> t
(.~) :: Traversal s t a b -> b -> s -> t