Reflex.Dom.Prerender

Documentation

class Prerender js m | m -> js where

Minimal complete definition

prerenderClientDict

Methods

prerenderClientDict :: Maybe (Dict (PrerenderClientConstraint js m))

prerender :: forall js m a. Prerender js m => m a -> (PrerenderClientConstraint js m => m a) -> m a

Draw one widget when prerendering (e.g. server-side) and another when the widget is fully instantiated. In a given execution of this function, there will be exactly one invocation of exactly one of the arguments.