Reflex.Dom.Widget.Basic

Documentation

partitionMapBySetLT :: forall k v. Ord k => Set k -> Map k v -> Map (Either k ()) (Map k v)

Breaks the given Map into pieces based on the given Set. Each piece will contain only keys that are less than the key of the piece, and greater than or equal to the key of the piece with the next-smaller key. There will be one additional piece containing all keys from the original Map that are larger or equal to the largest key in the Set. Either k () is used instead of Maybe k so that the resulting map of pieces is sorted so that the additional piece has the largest key. No empty pieces will be included in the output.

listHoldWithKey :: forall t m k v a. (Ord k, DomBuilder t m, MonadHold t m) => Map k v -> Event t (Map k (Maybe v)) -> (k -> v -> m a) -> m (Dynamic t (Map k a))

newtype ChildResult t k a

Constructors

ChildResult

Fields

unChildResult :: (a, Event t (Map k (Maybe (ChildResult t k a))))

Displaying Values

text :: DomBuilder t m => Text -> m ()

dynText :: forall t m. (PostBuild t m, DomBuilder t m) => Dynamic t Text -> m ()

display :: (PostBuild t m, DomBuilder t m, Show a) => Dynamic t a -> m ()

button :: DomBuilder t m => Text -> m (Event t ())

dyn :: (DomBuilder t m, PostBuild t m) => Dynamic t (m a) -> m (Event t a)

Given a Dynamic of widget-creating actions, create a widget that is recreated whenever the Dynamic updates. The returned Event of widget results occurs when the Dynamic does. Note: Often, the type a is an Event, in which case the return value is an Event-of-Events that would typically be flattened (via switchPromptly).

widgetHold :: (DomBuilder t m, MonadHold t m) => m a -> Event t (m a) -> m (Dynamic t a)

Given an initial widget and an Event of widget-creating actions, create a widget that is recreated whenever the Event fires. The returned Dynamic of widget results occurs when the Event does. Note: Often, the type a is an Event, in which case the return value is a Dynamic-of-Events that would typically be flattened.

untilReady :: (DomBuilder t m, PostBuild t m) => m a -> m b -> m (a, Event t b)

Render a placeholder widget to be shown while another widget is not yet done rendering

Working with Maps

diffMapNoEq :: Ord k => Map k v -> Map k v -> Map k (Maybe v)

diffMap :: (Ord k, Eq v) => Map k v -> Map k v -> Map k (Maybe v)

applyMap :: Ord k => Map k (Maybe v) -> Map k v -> Map k v

mapPartitionEithers :: Map k (Either a b) -> (Map k a, Map k b)

applyMapKeysSet :: Ord k => Map k (Maybe v) -> Set k -> Set k

Apply a map patch to a set > applyMapKeysSet patch (Map.keysSet m) == Map.keysSet (applyMap patch m)

Widgets on Collections

listWithKey :: forall t k v m a. (Ord k, DomBuilder t m, PostBuild t m, MonadFix m, MonadHold t m) => Dynamic t (Map k v) -> (k -> Dynamic t v -> m a) -> m (Dynamic t (Map k a))

listWithKey' :: (Ord k, DomBuilder t m, MonadFix m, MonadHold t m) => Map k v -> Event t (Map k (Maybe v)) -> (k -> v -> Event t v -> m a) -> m (Dynamic t (Map k a))

Deprecated: listWithKey' has been renamed to listWithKeyShallowDiff; also, its behavior has changed to fix a bug where children were always rebuilt (never updated)

listWithKeyShallowDiff :: (Ord k, DomBuilder t m, MonadFix m, MonadHold t m) => Map k v -> Event t (Map k (Maybe v)) -> (k -> v -> Event t v -> m a) -> m (Dynamic t (Map k a))

Display the given map of items (in key order) using the builder function provided, and update it with the given event. Nothing update entries will delete the corresponding children, and Just entries will create them if they do not exist or send an update event to them if they do.

listViewWithKey :: (Ord k, DomBuilder t m, PostBuild t m, MonadHold t m, MonadFix m) => Dynamic t (Map k v) -> (k -> Dynamic t v -> m (Event t a)) -> m (Event t (Map k a))

Create a dynamically-changing set of Event-valued widgets. This is like listWithKey, specialized for widgets returning (Event t a). listWithKey would return 'Dynamic t (Map k (Event t a))' in this scenario, but listViewWithKey flattens this to 'Event t (Map k a)' via switch.

selectViewListWithKey

Arguments

:: (DomBuilder t m, Ord k, PostBuild t m, MonadHold t m, MonadFix m)

=> Dynamic t k
Current selection key

-> Dynamic t (Map k v)
Dynamic key/value map

-> (k -> Dynamic t v -> Dynamic t Bool -> m (Event t a))
Function to create a widget for a given key from Dynamic value and Dynamic Bool indicating if this widget is currently selected

-> m (Event t (k, a))
Event that fires when any child's return Event fires. Contains key of an arbitrary firing widget.

Create a dynamically-changing set of widgets, one of which is selected at any time.

selectViewListWithKey_

Arguments

:: (DomBuilder t m, Ord k, PostBuild t m, MonadHold t m, MonadFix m)

=> Dynamic t k
Current selection key

-> Dynamic t (Map k v)
Dynamic key/value map

-> (k -> Dynamic t v -> Dynamic t Bool -> m (Event t a))
Function to create a widget for a given key from Dynamic value and Dynamic Bool indicating if this widget is currently selected

-> m (Event t k)
Event that fires when any child's return Event fires. Contains key of an arbitrary firing widget.

Creating DOM Elements

el :: forall t m a. DomBuilder t m => Text -> m a -> m a

Create a DOM element > el "div" (text "Hello World") divHello World/div

elAttr :: forall t m a. DomBuilder t m => Text -> Map Text Text -> m a -> m a

Create a DOM element with attributes > elAttr "a" ("href" =: "http://google.com") (text "Google!") href="http://google.com"Google!/a

elClass :: forall t m a. DomBuilder t m => Text -> Text -> m a -> m a

Create a DOM element with classes > elClass "div" "row" (return ()) class="row"/div

elDynAttr :: forall t m a. (DomBuilder t m, PostBuild t m) => Text -> Dynamic t (Map Text Text) -> m a -> m a

Create a DOM element with Dynamic Attributes > elClass "div" (constDyn ("class" =: "row")) (return ()) class="row"/div

elDynClass :: forall t m a. (DomBuilder t m, PostBuild t m) => Text -> Dynamic t Text -> m a -> m a

Create a DOM element with a Dynamic Class > elDynClass "div" (constDyn "row") (return ()) class="row"/div

With Element Results

el' :: forall t m a. DomBuilder t m => Text -> m a -> m (Element EventResult (DomBuilderSpace m) t, a)

Create a DOM element and return the element > do (e, _) <- el' "div" (text Click) > return $ domEvent Click e

elAttr' :: forall t m a. DomBuilder t m => Text -> Map Text Text -> m a -> m (Element EventResult (DomBuilderSpace m) t, a)

Create a DOM element with attributes and return the element

elClass' :: forall t m a. DomBuilder t m => Text -> Text -> m a -> m (Element EventResult (DomBuilderSpace m) t, a)

Create a DOM element with a class and return the element

elDynAttr' :: forall t m a. (DomBuilder t m, PostBuild t m) => Text -> Dynamic t (Map Text Text) -> m a -> m (Element EventResult (DomBuilderSpace m) t, a)

Create a DOM element with Dynamic Attributes and return the element

elDynClass' :: forall t m a. (DomBuilder t m, PostBuild t m) => Text -> Dynamic t Text -> m a -> m (Element EventResult (DomBuilderSpace m) t, a)

Create a DOM element with a Dynamic class and return the element

elDynAttrNS' :: forall t m a. (DomBuilder t m, PostBuild t m) => Maybe Text -> Text -> Dynamic t (Map Text Text) -> m a -> m (Element EventResult (DomBuilderSpace m) t, a)

dynamicAttributesToModifyAttributes :: (Ord k, PostBuild t m) => Dynamic t (Map k Text) -> m (Event t (Map k (Maybe Text)))

List Utils

list :: (Ord k, DomBuilder t m, MonadHold t m, PostBuild t m, MonadFix m) => Dynamic t (Map k v) -> (Dynamic t v -> m a) -> m (Dynamic t (Map k a))

Create a dynamically-changing set of widgets from a Dynamic key/value map. Unlike the withKey variants, the child widgets are insensitive to which key they're associated with.

simpleList :: (DomBuilder t m, MonadHold t m, PostBuild t m, MonadFix m) => Dynamic t [v] -> (Dynamic t v -> m a) -> m (Dynamic t [a])

Create a dynamically-changing set of widgets from a Dynamic list.

Specific DOM Elements

newtype Link t

Constructors

Link

Fields

_link_clicked :: Event t ()

linkClass :: DomBuilder t m => Text -> Text -> m (Link t)

link :: DomBuilder t m => Text -> m (Link t)

divClass :: forall t m a. DomBuilder t m => Text -> m a -> m a

dtdd :: forall t m a. DomBuilder t m => Text -> m a -> m a

blank :: forall m. Monad m => m ()

Workflows

newtype Workflow t m a

Constructors

Workflow

Fields

unWorkflow :: m (a, Event t (Workflow t m a))

workflow :: forall t m a. (DomBuilder t m, MonadFix m, MonadHold t m) => Workflow t m a -> m (Dynamic t a)

workflowView :: forall t m a. (DomBuilder t m, MonadFix m, MonadHold t m, PostBuild t m) => Workflow t m a -> m (Event t a)

mapWorkflow :: DomBuilder t m => (a -> b) -> Workflow t m a -> Workflow t m b

mapWorkflowCheap :: DomBuilder t m => (a -> b) -> Workflow t m a -> Workflow t m b

Tables and Lists

tableDynAttr

Arguments

:: (Ord k, DomBuilder t m, MonadHold t m, PostBuild t m, MonadFix m)

=> Text
Class applied to table element

-> [(Text, k -> Dynamic t r -> m v)]
Columns of (header, row key -> row value -> child widget)

-> Dynamic t (Map k r)
Map from row key to row value

-> (k -> m (Dynamic t (Map Text Text)))
Function to compute tr element attributes from row key

-> m (Dynamic t (Map k (Element EventResult (DomBuilderSpace m) t, [v])))
Map from row key to (El, list of widget return values)

A widget to display a table with static columns and dynamic rows.

tabDisplay

Arguments

:: (MonadFix m, DomBuilder t m, MonadHold t m, PostBuild t m, Ord k)

=> Text
Class applied to ul element

-> Text
Class applied to currently active li element

-> Map k (Text, m ())
Map from (arbitrary) key to (tab label, child widget)

-> m ()

A widget to construct a tabbed view that shows only one of its child widgets at a time. Creates a header bar containing a ul with one li per child; clicking a li displays the corresponding child and hides all others.

class HasAttributes a where

Minimal complete definition

attributes

Associated Types

type Attrs a :: *

Methods

attributes :: Lens' a (Attrs a)